+++ src/test/java/io/ionic/libs/iongeolocationlib/controller/IONGLOCControllerTest.kt
@@ -3,14 +3,22 @@ package io.ionic.libs.iongeolocationlib.controller
 import android.app.Activity
 import android.app.PendingIntent
 import android.location.Location
+import android.location.LocationManager
+import android.net.ConnectivityManager
 import android.os.Build
+import android.os.CancellationSignal
 import android.os.Looper
 import android.util.Log
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.IntentSenderRequest
+import androidx.core.location.LocationListenerCompat
+import androidx.core.location.LocationManagerCompat
+import androidx.core.location.LocationRequestCompat
+import androidx.core.util.Consumer
 import app.cash.turbine.test
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
+import com.google.android.gms.common.api.ApiException
 import com.google.android.gms.common.api.ResolvableApiException
 import com.google.android.gms.common.api.Status
 import com.google.android.gms.location.CurrentLocationRequest
@@ -22,24 +30,34 @@ import com.google.android.gms.location.LocationSettingsResponse
 import com.google.android.gms.location.LocationSettingsResult
 import com.google.android.gms.location.SettingsClient
 import com.google.android.gms.tasks.Task
+import io.ionic.libs.iongeolocationlib.controller.helper.IONGLOCBuildConfig
+import io.ionic.libs.iongeolocationlib.controller.helper.IONGLOCFallbackHelper
+import io.ionic.libs.iongeolocationlib.controller.helper.IONGLOCGoogleServicesHelper
 import io.ionic.libs.iongeolocationlib.model.IONGLOCException
 import io.ionic.libs.iongeolocationlib.model.IONGLOCLocationOptions
 import io.ionic.libs.iongeolocationlib.model.IONGLOCLocationResult
 import io.mockk.coEvery
+import io.mockk.coVerify
 import io.mockk.every
+import io.mockk.just
 import io.mockk.mockk
 import io.mockk.mockkObject
 import io.mockk.mockkStatic
+import io.mockk.runs
+import io.mockk.slot
 import io.mockk.spyk
 import io.mockk.unmockkObject
 import io.mockk.unmockkStatic
 import io.mockk.verify
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.async
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.tasks.await
+import kotlinx.coroutines.test.advanceTimeBy
 import kotlinx.coroutines.test.advanceUntilIdle
+import kotlinx.coroutines.test.runCurrent
 import kotlinx.coroutines.test.runTest
 import org.junit.After
 import org.junit.Assert.assertEquals
@@ -56,9 +74,17 @@ class IONGLOCControllerTest {
     private val activityResultLauncher = mockk<ActivityResultLauncher<IntentSenderRequest>>()
     private val googleApiAvailability = mockk<GoogleApiAvailability>()
     private val locationSettingsClient = mockk<SettingsClient>()
-    private val helper = spyk(
-        IONGLOCServiceHelper(fusedLocationProviderClient, activityResultLauncher)
+    private val locationManager = mockk<LocationManager>()
+    private val connectivityManager = mockk<ConnectivityManager>()
+    private val googleServicesHelper = spyk(
+        IONGLOCGoogleServicesHelper(
+            locationManager,
+            connectivityManager,
+            fusedLocationProviderClient,
+            activityResultLauncher
+        )
     )
+    private val fallbackHelper = spyk(IONGLOCFallbackHelper(locationManager, connectivityManager))
 
     private val mockAndroidLocation = mockkLocation()
     private val locationSettingsTask = mockk<Task<LocationSettingsResponse>>(relaxed = true)
@@ -67,6 +93,7 @@ class IONGLOCControllerTest {
 
     private lateinit var sut: IONGLOCController
     private lateinit var locationCallback: LocationCallback
+    private lateinit var locationListenerCompat: LocationListenerCompat
 
     @Before
     fun setUp() {
@@ -82,16 +109,21 @@ class IONGLOCControllerTest {
         every { Log.d(any(), any(), any()) } returns 0
         mockkStatic(Looper::class)
         every { Looper.getMainLooper() } returns mockk<Looper>()
+        mockkStatic(LocationManagerCompat::class)
 
         sut = IONGLOCController(
             fusedLocationClient = fusedLocationProviderClient,
+            locationManager = locationManager,
+            connectivityManager = connectivityManager,
             activityLauncher = activityResultLauncher,
-            helper = helper
+            googleServicesHelper = googleServicesHelper,
+            fallbackHelper = fallbackHelper
         )
     }
 
     @After
     fun tearDown() {
+        unmockkStatic(LocationManagerCompat::class)
         unmockkStatic(Looper::class)
         unmockkStatic(Log::class)
         unmockkObject(IONGLOCBuildConfig)
@@ -218,13 +250,13 @@ class IONGLOCControllerTest {
 
             sut.addWatch(mockk<Activity>(), locationOptions, "1").test {
                 advanceUntilIdle()  // to wait until locationCallback is instantiated
-                emitLocations(listOf(mockAndroidLocation))
+                emitLocationsGMS(listOf(mockAndroidLocation))
                 var result = awaitItem()
                 assertTrue(result.isSuccess)
                 assertEquals(listOf(locationResult), result.getOrNull())
 
 
-                emitLocations(
+                emitLocationsGMS(
                     listOf(
                         mockkLocation { every { time } returns 1234L },
                         mockkLocation { every { time } returns 12345L },
@@ -268,7 +300,7 @@ class IONGLOCControllerTest {
 
             sut.addWatch(mockk<Activity>(), locationOptions, "1").test {
                 advanceUntilIdle()  // to wait until locationCallback is instantiated
-                emitLocations(listOf(mockAndroidLocation))
+                emitLocationsGMS(listOf(mockAndroidLocation))
                 val result = awaitItem()
 
                 assertTrue(result.isSuccess)
@@ -354,13 +386,210 @@ class IONGLOCControllerTest {
             sut.addWatch(mockk<Activity>(), locationOptions, watchId).test {
                 advanceUntilIdle()  // to wait until locationCallback is instantiated
 
-                emitLocations(listOf(mockAndroidLocation))
+                emitLocationsGMS(listOf(mockAndroidLocation))
 
                 ensureAllEventsConsumed()
             }
         }
     // endregion clearWatch tests
 
+    // region fallback tests
+    @Test
+    fun `given location settings check fails but enableLocationManagerFallback=true and there is cached data, when getCurrentLocation is called, result is returned`() =
+        runTest {
+            givenSuccessConditions() // to instantiate mocks
+            coEvery { locationSettingsTask.await() } throws RuntimeException()
+            val currentTime = System.currentTimeMillis()
+            every { locationManager.getLastKnownLocation(any()) } returns mockkLocation {
+                every { time } returns currentTime
+            }
+
+            val result = sut.getCurrentPosition(mockk<Activity>(), locationOptionsWithFallback)
+
+            assertTrue(result.isSuccess)
+            assertEquals(locationResult.copy(timestamp = currentTime), result.getOrNull())
+            coVerify(inverse = true) {
+                // only getLastKnownLocation, no location update requested
+                LocationManagerCompat.requestLocationUpdates(
+                    any(),
+                    any(),
+                    any<LocationRequestCompat>(),
+                    any(),
+                    any<Looper>()
+                )
+            }
+        }
+
+    @Test
+    fun `given location settings check fails with resolvableError but enableLocationManagerFallback=true but cached data is older, when getCurrentLocation is called, result is returned`() =
+        runTest {
+            givenSuccessConditions() // to instantiate mocks
+            givenResolvableApiException(Activity.RESULT_OK)
+            every { locationManager.getLastKnownLocation(any()) } returns mockkLocation {
+                every { time } returns System.currentTimeMillis() - (60_000L * 60_000L)
+            }
+
+            val deferred =
+                async { sut.getCurrentPosition(mockk<Activity>(), locationOptionsWithFallback) }
+            runCurrent() // to wait until locationListenerCompat is instantiated, can't use advanceUntilIdle because that would trigger the timeout
+            locationListenerCompat.onLocationChanged(mockAndroidLocation)
+            val result = deferred.await()
+
+            assertTrue(result.isSuccess)
+            assertEquals(locationResult, result.getOrNull())
+            coVerify {
+                // to confirm that listener has been removed by the end of getCurrentPosition
+                LocationManagerCompat.removeUpdates(locationManager, locationListenerCompat)
+            }
+            // to confirm that the correct quality was passed, based on the fact that
+            // 1. there is no network provider and 2. options#enableHighAccuracy=true
+            val slot = slot<LocationRequestCompat>()
+            coVerify {
+                // only getLastKnownLocation, no location update requested
+                LocationManagerCompat.requestLocationUpdates(
+                    any(),
+                    any(),
+                    capture(slot),
+                    any(),
+                    any<Looper>()
+                )
+            }
+            assertEquals(
+                LocationRequestCompat.QUALITY_BALANCED_POWER_ACCURACY,
+                slot.captured.quality
+            )
+        }
+
+    @Test
+    fun `given all preconditions pass and enableLocationManagerFallback=true, when getCurrentLocation is called, the fallback is not called`() =
+        runTest {
+            givenSuccessConditions() // to instantiate mocks
+
+            sut.getCurrentPosition(mockk<Activity>(), locationOptionsWithFallback)
+
+            coVerify(inverse = true) {
+                fallbackHelper.getCurrentLocation(any())
+            }
+        }
+
+    @Test
+    fun `given location settings check fails with resolvableError, location is off, and enableLocationManagerFallback=true, when getCurrentLocation is called, the fallback is not called`() =
+        runTest {
+            givenSuccessConditions() // to instantiate mocks
+            givenResolvableApiException(Activity.RESULT_OK)
+            every { LocationManagerCompat.isLocationEnabled(any()) } returns false
+
+            sut.getCurrentPosition(mockk<Activity>(), locationOptionsWithFallback)
+
+            coVerify(inverse = true) {
+                fallbackHelper.getCurrentLocation(any())
+            }
+        }
+
+    @Test
+    fun `given fallback is being used but requestLocationUpdates does not notify listener, when getCurrentLocation is called, IONGLOCLocationRetrievalTimeoutException is returned`() =
+        runTest {
+            givenSuccessConditions() // to instantiate mocks
+            coEvery { locationSettingsTask.await() } throws RuntimeException()
+            every { LocationManagerCompat.isLocationEnabled(any()) } returns false
+
+            val deferred =
+                async { sut.getCurrentPosition(mockk<Activity>(), locationOptionsWithFallback) }
+            advanceTimeBy(locationOptionsWithFallback.timeout * 2)
+            val result = deferred.await()
+
+            assertTrue(result.isFailure)
+            assertTrue(result.exceptionOrNull() is IONGLOCException.IONGLOCLocationRetrievalTimeoutException)
+        }
+
+    @Test
+    fun `given SETTINGS_CHANGE_UNAVAILABLE error and network+location disabled and enableLocationManagerFallback=true, when getCurrentLocation is called, IONGLOCLocationAndNetworkDisabledException is returned`() =
+        runTest {
+            givenSuccessConditions() // to instantiate mocks
+            coEvery { locationSettingsTask.await() } throws mockk<ApiException> {
+                every { message } returns "8502: SETTINGS_CHANGE_UNAVAILABLE"
+            }
+            every { LocationManagerCompat.isLocationEnabled(any()) } returns false
+
+            val result = sut.getCurrentPosition(mockk<Activity>(), locationOptionsWithFallback)
+
+            assertTrue(result.isFailure)
+            assertTrue(result.exceptionOrNull() is IONGLOCException.IONGLOCLocationAndNetworkDisabledException)
+        }
+
+    @Test
+    fun `given play services not available but enableLocationManagerFallback=true, when addWatch is called, locations returned in flow`() =
+        runTest {
+            givenSuccessConditions()
+            givenPlayServicesNotAvailableWithResolvableError()
+
+            sut.addWatch(mockk<Activity>(), locationOptionsWithFallback, "1").test {
+                advanceUntilIdle()  // to wait until locationListenerCompat is instantiated
+                emitLocationsFallback(listOf(mockAndroidLocation))
+                var result = awaitItem()
+                assertTrue(result.isSuccess)
+                assertEquals(listOf(locationResult), result.getOrNull())
+
+
+                emitLocationsFallback(
+                    listOf(
+                        mockkLocation { every { time } returns 1234L },
+                        mockkLocation { every { time } returns 12345L },
+                    )
+                )
+                result = awaitItem()
+                assertEquals(
+                    listOf(
+                        locationResult.copy(timestamp = 1234L),
+                        locationResult.copy(timestamp = 12345L),
+                    ),
+                    result.getOrNull()
+                )
+            }
+        }
+
+    @Test
+    fun `given play services not available but enableLocationManagerFallback=true and there is cached location, when addWatch is called, cached location returned in flow`() =
+        runTest {
+            givenSuccessConditions()
+            givenPlayServicesNotAvailableWithUnResolvableError()
+            val currentTime = System.currentTimeMillis()
+            every { locationManager.getLastKnownLocation(any()) } returns mockkLocation {
+                every { time } returns currentTime
+            }
+
+            sut.addWatch(mockk<Activity>(), locationOptionsWithFallback, "1").test {
+                advanceUntilIdle()  // to wait until locationListenerCompat is instantiated
+
+                val result = awaitItem()
+                assertTrue(result.isSuccess)
+                assertEquals(
+                    listOf(locationResult.copy(timestamp = currentTime)),
+                    result.getOrNull()
+                )
+                expectNoEvents()
+            }
+        }
+
+    @Test
+    fun `given watch was added via fallback, when clearWatch is called, true is returned`() =
+        runTest {
+            val watchId = "id"
+            givenSuccessConditions()
+            givenPlayServicesNotAvailableWithUnResolvableError()
+            sut.addWatch(mockk<Activity>(), locationOptionsWithFallback, watchId).test {
+                advanceUntilIdle()  // to wait until locationListenerCompat is instantiated
+
+                val result = sut.clearWatch(watchId)
+
+                assertTrue(result)
+                expectNoEvents()
+            }
+            verify { LocationManagerCompat.removeUpdates(any(), locationListenerCompat) }
+        }
+    // endregion fallback tests
+
+    // region utils
     private fun givenSuccessConditions() {
         every { googleApiAvailability.isGooglePlayServicesAvailable(any()) } returns ConnectionResult.SUCCESS
         every { locationSettingsClient.checkLocationSettings(any()) } returns locationSettingsTask
@@ -375,7 +604,6 @@ class IONGLOCControllerTest {
             fusedLocationProviderClient.getCurrentLocation(any<CurrentLocationRequest>(), any())
         } returns currentLocationTask
         coEvery { currentLocationTask.await() } returns mockAndroidLocation
-
         every {
             fusedLocationProviderClient.requestLocationUpdates(
                 any(),
@@ -386,8 +614,38 @@ class IONGLOCControllerTest {
             locationCallback = args[1] as LocationCallback
             voidTask
         }
-
         every { fusedLocationProviderClient.removeLocationUpdates(any<LocationCallback>()) } returns voidTask
+
+        every { connectivityManager.activeNetwork } returns null
+        every { LocationManagerCompat.hasProvider(any(), any()) } returns true
+        every { LocationManagerCompat.isLocationEnabled(any()) } returns true
+        every { locationManager.getLastKnownLocation(any()) } returns null
+        val consumerSlot = slot<Consumer<Location>>()
+        every {
+            LocationManagerCompat.getCurrentLocation(
+                any(),
+                any(),
+                any<CancellationSignal>(),
+                any(),
+                capture(consumerSlot)
+            )
+        } answers {
+            consumerSlot.captured.accept(mockAndroidLocation)
+        }
+        every {
+            LocationManagerCompat.requestLocationUpdates(
+                any(),
+                any(),
+                any<LocationRequestCompat>(),
+                any(),
+                any<Looper>()
+            )
+        } answers {
+            locationListenerCompat = args[3] as LocationListenerCompat
+        }
+        every {
+            LocationManagerCompat.removeUpdates(any(), any())
+        } just runs
     }
 
     private fun givenPlayServicesNotAvailableWithResolvableError() {
@@ -428,7 +686,7 @@ class IONGLOCControllerTest {
             overrideDefaultMocks()
         }
 
-    private fun emitLocations(locationList: List<Location>) {
+    private fun emitLocationsGMS(locationList: List<Location>) {
         locationCallback.onLocationResult(
             mockk<LocationResult>(relaxed = true) {
                 every { locations } returns locationList.toMutableList()
@@ -436,15 +694,28 @@ class IONGLOCControllerTest {
         )
     }
 
+    private fun emitLocationsFallback(locationList: List<Location>) {
+        if (locationList.size == 1) {
+            locationListenerCompat.onLocationChanged(locationList.first())
+        } else {
+            locationListenerCompat.onLocationChanged(locationList)
+        }
+    }
+    // endregion utils
+
     companion object {
         private const val DELAY = 3_000L
 
         private val locationOptions = IONGLOCLocationOptions(
-            timeout = 5000,
-            maximumAge = 3000,
+            timeout = 60_000,
+            maximumAge = 30_000,
             enableHighAccuracy = true,
-            minUpdateInterval = 2000L
+            minUpdateInterval = 2000L,
+            enableLocationManagerFallback = false
         )
+
+        private val locationOptionsWithFallback =
+            locationOptions.copy(enableLocationManagerFallback = true)
 
         private val locationResult = IONGLOCLocationResult(
             latitude = 1.0,