+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0]
+
+### 2025-09-30
+
+- Feature: Allow using a fallback if Google Play Services fails.
+
+BREAKING CHANGE: The constructor for the controller and some of its methods have changed signature.
+You will need to change how your application calls the library if you update to this version.
 
 ### 2025-06-26